changed code structure and fixed cell editing functionality
