--- src/scripts/modules/constants.js
@@ -1,4 +1,0 @@
-export const table = document.querySelector('table');
-export const tbody = table.querySelector('tbody');
-export const thead = table.querySelector('thead');
-export const theadRow = thead.querySelector('tr');
